requirements.txt
ibm-watsonx-orchestrate
requests
google-generativeai
selenium
webdriver-manager
pyyaml
python-dotenv
